requirementes.txt
Flask
Requests
mysql-connector-python